requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
pymongo==4.6.1
python-jose==3.3.0
gunicorn==21.2.0
pytest==7.4.3 